// src/core/http/generated/models/tenant.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { TenantStatus } from './tenantStatus.ts';

export interface Tenant {
  id: string;
  name: string;
  email: string;
  documento?: string;
  telefone?: string;
  status: TenantStatus;
  plano?: string;
  created_at: string;
  updated_at?: string;
}