// src/core/http/generated/models/admin-tenants/storeAdminTenantRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Tenants
 * Filtered OpenAPI specification for Admin Tenants.
 * OpenAPI spec version: 0.0.1
 */

export interface StoreAdminTenantRequest {
  /** @maxLength 255 */
  name: string;
  /** @maxLength 255 */
  slug: string;
  /**
     * @maxLength 255
     * @nullable
     */
  domain?: string | null;
  /**
     * @maxLength 255
     * @nullable
     */
  document?: string | null;
  /**
     * @maxLength 255
     * @nullable
     */
  email?: string | null;
  /**
     * @maxLength 255
     * @nullable
     */
  phone?: string | null;
  active?: boolean;
  /**
     * @maxLength 255
     * @nullable
     */
  plan?: string | null;
  /** @nullable */
  plan_id?: number | null;
  /** @nullable */
  trial_ends_at?: string | null;
}